Move reading caCert and token files out of the k8s library
